remove subsidy code from zebra-consensus, previously copied into zebra-chain
