+++ zebra-consensus/src/error.rs
@@ -9,7 +9,9 @@ use chrono::{DateTime, Utc};
 use thiserror::Error;
 
 use zebra_chain::{
-    amount, block, orchard, sapling, sprout,
+    amount, block, orchard,
+    parameters::subsidy::SubsidyError,
+    sapling, sprout,
     transparent::{self, MIN_TRANSPARENT_COINBASE_MATURITY},
 };
 use zebra_state::ValidateContextError;
@@ -21,35 +23,6 @@ use proptest_derive::Arbitrary;
 
 /// Workaround for format string identifier rules.
 const MAX_EXPIRY_HEIGHT: block::Height = block::Height::MAX_EXPIRY_HEIGHT;
-
-/// Block subsidy errors.
-#[derive(Error, Clone, Debug, PartialEq, Eq)]
-#[allow(missing_docs)]
-pub enum SubsidyError {
-    #[error("no coinbase transaction in block")]
-    NoCoinbase,
-
-    #[error("funding stream expected output not found")]
-    FundingStreamNotFound,
-
-    #[error("miner fees are invalid")]
-    InvalidMinerFees,
-
-    #[error("a sum of amounts overflowed")]
-    SumOverflow,
-
-    #[error("unsupported height")]
-    UnsupportedHeight,
-
-    #[error("invalid amount")]
-    InvalidAmount(amount::Error),
-}
-
-impl From<amount::Error> for SubsidyError {
-    fn from(amount: amount::Error) -> Self {
-        Self::InvalidAmount(amount)
-    }
-}
 
 /// Errors for semantic transaction validation.
 #[derive(Error, Clone, Debug, PartialEq, Eq)]
