+++ zebra-consensus/src/lib.rs
@@ -48,10 +48,7 @@ pub mod transaction;
 pub use block::check::difficulty_is_valid;
 
 pub use block::{
-    subsidy::{
-        funding_streams::{funding_stream_address, funding_stream_values, new_coinbase_script},
-        general::{block_subsidy, miner_subsidy},
-    },
+    subsidy::funding_streams::{funding_stream_address, new_coinbase_script},
     Request, VerifyBlockError, MAX_BLOCK_SIGOPS,
 };
 pub use checkpoint::{
